fix: predicted_goals changed to float

+++ backend/api/dto/team_dto.py
@@ -5,6 +5,6 @@ from matches.match import Match
 
 class team_dto(BaseModel):
     current_goals: int
-    predicted_goals: int
+    predicted_goals: float
     actual_goals: int
